initial work

+++ src/main/java/gregicality/science/common/pipelike/pressure/PressurePipeData.java
@@ -4,13 +4,13 @@ import gregicality.science.api.GCYSValues;
 
 public class PressurePipeData {
 
-    public static final PressurePipeData EMPTY = new PressurePipeData(GCYSValues.EARTH_PRESSURE, GCYSValues.EARTH_PRESSURE, 1.0);
+    public static final PressurePipeData EMPTY = new PressurePipeData(GCYSValues.EARTH_PRESSURE, GCYSValues.EARTH_PRESSURE, 1);
 
     private final double minPressure;
     private final double maxPressure;
-    private final double volume;
+    private final int volume;
 
-    public PressurePipeData(double minPressure, double maxPressure, double volume) {
+    public PressurePipeData(double minPressure, double maxPressure, int volume) {
         this.minPressure = minPressure;
         this.maxPressure = maxPressure;
         this.volume = volume;
@@ -24,7 +24,7 @@ public class PressurePipeData {
         return maxPressure;
     }
 
-    public double getVolume() {
+    public int getVolume() {
         return volume;
     }
 }
